swap recur with fn ast if no loop
